+++ src/drafts/MarkdownEditor/_Footer.tsx
@@ -1,13 +1,14 @@
-import React, {memo, useContext} from 'react'
-import {AlertIcon, ImageIcon, MarkdownIcon} from '@primer/octicons-react'
+import React, {memo} from 'react'
+import {AlertIcon} from '@primer/octicons-react'
 
-import {Spinner, Button, ButtonProps, LinkButton, Box, Text} from '../..'
-import {MarkdownEditorContext} from './_MarkdownEditorContext'
+import {Spinner, ButtonProps, Box, Text} from '../..'
+import {DefaultActionButtons} from './Actions'
 
 const uploadingNote = ([current, total]: [number, number]) =>
   total === 1 ? `Uploading your file...` : `Uploading your files... (${current}/${total})`
 
 export const Footer = ({
+  footerButtons,
   actionButtons,
   uploadButtonProps,
   fileUploadProgress,
@@ -15,6 +16,7 @@ export const Footer = ({
   errorMessage,
   previewMode,
 }: {
+  footerButtons: React.ReactNode
   actionButtons: React.ReactNode
   uploadButtonProps: Partial<ButtonProps> | null
   fileUploadProgress?: [number, number]
@@ -24,6 +26,7 @@ export const Footer = ({
 }) => (
   <Box sx={{pt: 2, display: 'flex', gap: 2, justifyContent: 'space-between', minHeight: '36px'}} as="footer">
     <Box sx={{display: 'flex', gap: 1, alignItems: 'center', fontSize: 0}}>
+      <Box sx={{display: 'flex', gap: 2}}>{footerButtons}</Box>
       {previewMode ? (
         <></>
       ) : fileUploadProgress ? (
@@ -33,16 +36,7 @@ export const Footer = ({
       ) : errorMessage ? (
         <ErrorMessage message={errorMessage} />
       ) : (
-        <>
-          <MarkdownSupportedHint />
-
-          {uploadButtonProps && (
-            <>
-              <VisualSeparator />
-              <FileUploadButton fileDraggedOver={fileDraggedOver} {...uploadButtonProps} />
-            </>
-          )}
-        </>
+        <DefaultActionButtons uploadButtonProps={uploadButtonProps} fileDraggedOver={fileDraggedOver} />
       )}
     </Box>
     {!fileUploadProgress && <Box sx={{display: 'flex', gap: 2}}>{actionButtons}</Box>}
@@ -57,50 +51,3 @@ const ErrorMessage = memo(({message}: {message: string}) => (
     {message}
   </Text>
 ))
-
-const FileUploadButton = memo(({fileDraggedOver, ...props}: Partial<ButtonProps> & {fileDraggedOver: boolean}) => {
-  const {condensed, disabled} = useContext(MarkdownEditorContext)
-
-  return (
-    <Button
-      variant="invisible"
-      leadingIcon={ImageIcon}
-      size="small"
-      sx={{color: 'fg.default', fontWeight: fileDraggedOver ? 'bold' : 'normal', px: 2}}
-      onMouseDown={(e: React.MouseEvent) => {
-        // Prevent pulling focus from the textarea
-        e.preventDefault()
-      }}
-      disabled={disabled}
-      {...props}
-    >
-      {condensed ? 'Add files' : fileDraggedOver ? 'Drop to add files' : 'Paste, drop, or click to add files'}
-    </Button>
-  )
-})
-
-const VisualSeparator = memo(() => (
-  <Box sx={{borderRightStyle: 'solid', borderRightWidth: 1, borderRightColor: 'border.muted', height: '100%'}} />
-))
-
-const MarkdownSupportedHint = memo(() => {
-  const {condensed} = useContext(MarkdownEditorContext)
-
-  return (
-    <LinkButton
-      leadingIcon={MarkdownIcon}
-      variant="invisible"
-      size="small"
-      sx={{color: 'inherit', fontWeight: 'normal', px: 2}}
-      href="https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax"
-      target="_blank"
-      // The markdown editor aria-description already describes it as Markdown editor, so it's
-      // redundant to say Markdown is supported again here. However for sighted users, they
-      // cannot see the aria-description so this is a useful hint for them. So the aria-label
-      // is different from the visible text content.
-      aria-label="Markdown documentation"
-    >
-      {!condensed && <Text aria-hidden>Markdown is supported</Text>}
-    </LinkButton>
-  )
-})
